test: updates reviewAgreement test and moves create agreement test to new module
